test_arize_set_attributes
